wmf: Provide a privileged (root-ful) mode image
With rootless mode, there is less isolation between the build container
processes that the runc worker spawns. The build containers also share
the same network namespace which leads to collisions in port allocation,
etc.

We're experimenting with running buildkitd in privileged mode for better
process isolation and using its support for CNI for better network
isolation. See https://www.cni.dev for information about the latter.

Define a Dockerfile stage for the privileged runtime and publish both
the rootless and privileged variants.

Bug: T329213