buildkitd: Add ability to restrict frontends and gateway sources
When building images on trusted runners we want to ensure that only
the Blubber frontend is used.  This commit adds the ability to
restrict the frontends and gateway sources in buildkitd.

buildkit.toml entries that will be used on trusted runners:

allowed-frontends = ["gateway.v0"]
allowed-gateway-sources = ["docker-registry.wikimedia.org/repos/releng/blubber/buildkit"]

Bug: T329220